Make it possible to have interactive debug views.
Sometimes we'd like to visualize something beyond an image in a debugging
view. A good example would be a vector field. While it may be possible to
represent vector directions with colors, it's not perfect and it gets even
harder to represent vector magnitudes with a large range of values.

This commit introduces the infrastructure to provide a factory producing
a custom widget for a particular debug view instead of providing an image
like it was done before. Arguments to the factory may be images or other
data, and as before those are not stored in memory but on disk and then
loaded and unloaded on demand.