Require Visual Studio 2013 on Windows.
Historically, both Visual Studio and MinGW were supported for building
on Windows. One good reason for that was the fact that 64-bit compilers
weren't included in the free Express Edition of Visual Studio. Since
that's no longer the case, we are dropping support for MinGW in order
to cut down maintenance efforts.

Additionally, when building on Windows, propagate some locations
from "Scan Tailor Dependencies" build to the main one. This should
reduce chances of CMake picking up the wrong versions of things like
boost and Qt.