Get rid of X11-specific code.
With Qt4 when running on X11 we had to make direct calls to XRender
in order to achieve acceptable performance when drawing large pixmaps
with affine transformation. There were a few other places where we had
to make special arrangements when running on X11.

With Qt5 none of that is necessary, as the X11 paint engine is gone
and the raster paint engine is used instead, just the way it was with
Qt4 on Windows.

Making direct XRender calls is now not just unnecessary but also
impossible, as QPixmap is no longer bound to an XRender Picture object.
Fortunately, the raster paint engine seems to deliver acceptable
performance, at least when running in VirtualBox (haven't tried natively).

This commit removes all code dependent on Q_WS_X11 preprocessor definition
(which Qt5 lacks anyway) and removes a direct dependency on libxrender.