Introduce the option to scale the output image
When working with low resolution input images it often makes sense
to upscale them as a processing step, in particular when binarization
would follow.

Previously, such upscaling was achieved by specifying the output DPI
to be higer than the input one. Since the output DPI option was removed
as part of major architecture change, a new upscaling mechanism was
required.

This commit introduces 3 buttons on the output stage to perform 1x,
(no upscaling) 1.5x and 2x upscaling.