Improve dewarping rasterization quality
The existing code didn't do a wery good job when dewarping was combined
with upscaling. We used to have the exact same problem with affine
transformations, which was solved by parametrising the area mapping
algorithm with the minimum rectangle of the source image to be mapped
to a single pixel in the destination image. When these rectangles are
unconstrained and are significantly smaller than a unit square, the
result looks like a nearest-neighbour algorithm. When constrained to
be at least 1.0 x 1.0 pixels, the result when upscaling is very close
to what we would get from a billinear transform, while still getting
all the benefits of the area mapping algorithm when downscaling.

This commit makes the minumum mapping area configurable and sets the
default to 0.9 x 0.9 px, which will result in slightly less smooting
than a billinear transform when upscaling, and pre-existing behaviour
when downscaling.