Provide move semantics to Grid template class
While at it, a custom replacement for std::swap() was replaced by
std::swap() itself. The original justification for a custom replacement
was the reluctance to include a heavy <algorithm> header to Grid.h.
This is no longer a concern with C++11, as std::swap() is declared
in <utility> header now.