Improve accuracy of anisotropic gaussian blur
In [1] they decompose a 2D gaussian into a product of a horizontal 1D
gaussian and an oriented gaussian with a zero standard deviation in one
of the directions. That approach produces the largest errors when
decomposing a nearly-vertical gaussian. We extend the approach of [1]
by choosing whether to do a horizontal or a vertical decomposition.
The new approach achieves lower errors overall, while the largest
errors now correspond to diagonal gaussians.

This commit also adds tests for isotropic and anisotropic gaussian blur.

[1]: Geusebroek, J-M., Arnold WM Smeulders, and Joost Van De Weijer.
"Fast anisotropic gauss filtering." Image Processing, IEEE Transactions
on 12.8 (2003): 938-943.