Initial support for OpenCL
Right now only one operation is accelerated: the filter bank for text
line detection (part of dewarping and de-keystoning operations). In this
case however, OpenCL helps only slightly, as the task is memory bandwidth
constrained.

As a side effect of this change, minor changes were also made to OpenGL
support and to loading translations.