Penalize dewarped lines for not being horizontal
When picking two horizontal curves to use for dewarping, we use the
Ransac algorithm. It samples random pairs of curves, build a distortion
model based on those, and dewarps the other candidate curves using that
model. The pair that gives the best dewarping on other candidate lines
wins.

Previously, we only considered how straight the dewarped lines are when
evaluating the quality of dewarping. This commit changes that to also
consider the slope of a dewarped line (the more horizontal the better).