Fix a Qt runtime warning
The warning was about size constraint violations on the thumbnail list
widget. It was caused by conflicting minimum and maximum widths, namely
maximum being smaller than the minimum. The maximum was specified in a
*.ui file while the minimum was computed and set at run time.

The fix is to remove the maximum width specification from the *.ui file
and set fixed rather than minimum width at run time.