Don't go to first page on the Output stage when output is not yet possible
If the user goes to the Output stage when output sizes for some of the
pages are unknown, we display an error message and we also used to go
to the first page at that point. The reasoning was: the user will likely
start batch processing when output is finally allowed and it's better
to start batch processing from the first page to process all of them.
Nonetheless, this behaviour was confusing and this commit removes it.