Change the way pages without a content box are handled
Previously, a small fake content box was being created for pages without
one. This was done to avoid introducing a special case in processing
code. For the user it was confusing though, and it got even more confusing
due to the changes made when moving dewarping to the Deskew stage.

This change stops creation of fake content boxes for pages without one,
introducing a special case instead. Such pages are outputted as pure
white images of size matching the other pages, as if appropriate option
was selected on the Page Layout stage.