Prepare downscaled images for displaying purposes in the background thread
ImageViewBase operates on two images: a full size one and a downscaled
one. The downscaled one is used for all real-time rendering while the
full-size one is used for creating a high-quality rendering once things
settle down. ImageViewBase can either create the downscaled image itself
or take the one provided by client code.

The problem with letting ImageViewBase create the downscaled image by
itself is that it happens in the GUI thread, blocking the GUI for a
moment.

This commit adds code to create downscaled images to be passed to
ImageViewBase while still in a background thread.