Make the existing filter bank generate the flow direction map
We need a direction map that for each pixel indicates the direction
of a nearby text line. We used to have a dedicated algorithm for
computing such a map, however it can be also generated as a side effect
of applying a filter bank for text line detection purpuses.

This commit re-purposes the filter bank to produce such a map and
removes the dedicated algorithm for doing so. While there, the vertical
bounds detection code was factored out into a separate file.