Fix synchonisation problems with OpenCL
The common two lines following the launch of a kernel were:

deps.clear();
deps.push_back(std::move(evt));

Unfortunately that std::move() makes it very error-prone, as we need
to remember not to do that before evt.wait() or the following construct
typically appearing at the end of a function:

if (event) {
    *event = evt;
}

In a few cases we actually failed to get it right. The solution is to
stop using std::move() for OpenCL events.