Simplify the OpenCL implementation of Gaussian blur
Switch the OpenCL implementation of Gaussian blur to a normalized scheme,
where every pass keeps the scale of the data. This requires an addition
multiplication in the inner loop of every other pass, but it does make
the implementation much simpler.

In addition to that, we no longer compute the history update matrix
in the kernel. Instead, we pre-compute it and pass it as a parameter
to the kernel.