Constrain the area of the image we apply dewarping to
Originally, the idea was to make the dewarped image at the Select Content
stage contain all the pixels in the crop area of original warped image.
Unfortunately it turns out same areas of the warped image may be unsafe
to dewarp. For instance, in the presence of perspective distortion, a
vanishing point (if it's within the image) can't be mapped to dewarped
coordinates.

The solution is to constrain the area of the original image that will
be dewarped to exlude unsafe areas.