Fix a crash on switching to "Match size by scaling"
The crash could be reproduced as follows:

1. Go to the Margins stage.
2. Select "Don't match size".
3. Apply to ... -> All Pages.
4. Select a different page.
5. Select "Match size by scaling" -> Crash

The crash itself seems to be a bug in Qt triggered by trying to draw
an empty rectangle with QPainter. However, such a situation was triggered
by a bug in Scan Tailor, which is addressed by this commit. The issue
was a missing update of m_aggregateHardSize and m_committedAggregateHardSize
in page_layout::ImageView::matchSizeModeChanged(). Indeed, when all pages
are set to "Don't match size", no pages participate in aggregate hard
size calculation, leading aggregate hard size to be "0.0 x 0.0". That's
not a problem when in "Don't match size" mode but is a problem in other
modes.