Fix crash for some scans without content margins
For some scans app  crashes on line 291.

The problem must be with rounding rectangles in float coordinates to rectangles in integer coordinates. In some circumstances it rounds to -1 pixel for output rect and to +1 pixel for working rect. Working rect size becomes 1 pixel bigger than output page size and everything collapses. And non-zero margins seems to save ST from that. toAlignedRect() is using ceil() instead of round() for float coordinates conversion so the sizes should always match.
UPD: Sometimes `toAlignedRect` may return rect with coordinate < 0.0. Made additional checks for that.