Dewarp: reset distortion model from context menu
Context menu with one action is added for Dewarping view tab. It allows to reset current distortion model.
Resolves https://github.com/scantailor/scantailor/issues/190

"Automatic dewarping can be very useful. But sometimes it can lead to a messy mesh that is also complex and hard to straighten. There should be an option to reset the mesh to the square/straight."