Fix Auto dewarping for pages with undefined distortion model
There was a check for empty distortion model in ST, that resets it to default one.
That affects pages without text where distortion model can't be defined.
The check was saving the app from crash due to access to non existing model lines.
After commit "Scantailor Featured - Auto_Dewarping_Vert_Half_Correction " that safity check
was shifted after block of code that actually needs it.
So for some pages the app crash if Auto-dewarp was on.
I've moved this check back to the right place.