Improve invalidate Thumbnail perforance
Since ST Enhanced commit "thumbnails can be aligned horizontally if widget size changes" the
original implementation of ThumbnailSequence::Impl::invalidateThumbnailImpl(ItemsById::iterator const id_it);
has been raplaced with a simple call of invalidateAllThumbnails();
that dramatically hit perforance, especially in "Apply to" dialogs on a last step.

I've rollback the code of this function and made some changes in it to support multicolumn thumbnails.
Hope the required modifications of original implementation are correct.