Insert empty pages from context menu
Now one can add an empty pages to the project by inserting them above or below the selected page in
its context menu. The empty page is actually a 72 dpi A10 png image encoded in the app resources.
It's not created in source images folder but used from resources system directly.
There is a tricky code to make sure the empty page will be saved with a right filename in out dir.
This is required to make sure the order of empty pages in out folder is the same as natural order
of pages in the application. This allows to just select all files in out folder while importing in
DjVu encoders like DjVu Small and be sure their alphabetical order the same as pages order.

FileNameDisambiguator now keeps page_num info as well as filename and no suffix to output filename
is added in case source image filepath leads to the app resource system. This is bcs internally
all empty pages are represented as different pages of single multipage file.
It's not really a multipage image - it's just simulated.