[Exp] Copy zones from one page to others
This is a new command in zone context menu that allows you to copy
it to another pages. Default pages selection dialog is used. Fill
and picture zones are affected.

Technically this creates a new zone on selected pages with virtual
coordinates, and id. Just before page is processed all its zones
are checked for virtual flag. If virtual zone is found its
coordinates are transformed to current page coordinates and the
flag is reseted. This is required for splitted pages etc.
Also if zone was copied to page (has virtual flag regardless of
its state). A new context menu command "Delete from.." is
displayed. This allows to delete this zone from all selected pages
by its id.