Bugfix: fix TIFFTAG_FILLORDER tag in output tiff files
According to TIFF spec rev. 6
"FillOrder = 2 should be used only when BitsPerSample = 1 and the data is
either uncompressed or compressed using CCITT 1D or 2D compression, to
avoid potentially ambigous situations.
Support for FillOrder=2 is not required in a Baseline TIFF compliant reader
Default is FillOrder = 1."

FillOrder=2 in some files is suspected to cause problems when used with
an old non-libtiff based readers.
Like 3rd party Delphi component used in Scan Kromsator.