Basic app styling support
This will allow to change app style and apply stylesheets to it.
The settings are available on general page of settings menu.

(In fact any Qt app can be styled with `-style` and `-stylesheet`
keys in command line so this is just a UI way to do the same.
Also some hardcoded colors in the app is disabled to make it more
customizable with stylesheets.)

App will automatically list all available Qt styles found in system.
Stylesheets are deployed with the App. By default there is ubuntu.qss
from https://github.com/GTRONICK/QSS/blob/master/Ubuntu.qss
and my fork of BreezeStyleSheets:
https://github.com/trufanov-nok/BreezeStyleSheets-for-STU
To get the last one you should clone the STU project with
`git clone --recurse-submodules https://github.com/trufanov-nok/scantailor-universal`
as it's linked to the project as a submodule.

qss files are stored openly in filesystem to encourage user for
their editing and making own stylesheets.

Currently only linux deployment of qss is tuned.
Win support will be enabled later.