Copy source image ICC profile (off by default)
This patch adds exiv2 library usage to copy embedded ICC Profile (if exists)
from the source image (any format) to the resulting TIFF image.
This feature is off by default and must be switched on in settings.
As some ICC profiles doesn't support some color spaces the colorspace
optimisation (RGB image saved as grayscale if all pixels are shades of gray)
is disabled if the feature is on.